Prevent creation and deletion of Fleet-reserved profile names

+++ server/service/mdm.go
@@ -24,6 +24,7 @@ import (
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/mdm"
 	apple_mdm "github.com/fleetdm/fleet/v4/server/mdm/apple"
+	microsoft_mdm "github.com/fleetdm/fleet/v4/server/mdm/microsoft"
 	"github.com/fleetdm/fleet/v4/server/ptr"
 	"github.com/go-kit/kit/log/level"
 	"github.com/go-sql-driver/mysql"
@@ -1116,6 +1117,12 @@ func (svc *Service) DeleteMDMWindowsConfigProfile(ctx context.Context, profileUU
 		return ctxerr.Wrap(ctx, err)
 	}
 
+	// prevent deleting Windows OS Updates profile (controlled by the OS Updates settings)
+	if _, ok := microsoft_mdm.FleetReservedProfileNames()[prof.Name]; ok {
+		err := &fleet.BadRequestError{Message: fmt.Sprintf("Profile name %q is reserved, it cannot be deleted.", prof.Name)}
+		return ctxerr.Wrap(ctx, err, "validate profile")
+	}
+
 	if err := svc.ds.DeleteMDMWindowsConfigProfile(ctx, profileUUID); err != nil {
 		return ctxerr.Wrap(ctx, err)
 	}
@@ -1276,6 +1283,8 @@ func (svc *Service) NewMDMWindowsConfigProfile(ctx context.Context, teamID uint,
 			InternalErr: err,
 		})
 	}
+
+	// TODO(mna) : prevent creating a profile with a Fleet-reserved name.
 
 	cp := fleet.MDMWindowsConfigProfile{
 		TeamID: &teamID,
